# docs/openapi.yaml
openapi: 3.0.3
info:
  title: go-file AI API
  description: >-
    机器友好的只读发现 + 上传/下载接口，供 AI 代理获取文件列表、元数据、统计，并自主下载或回传文件。
    所有 /api/ai/* 接口（manifest 除外）需在 Authorization 请求头携带用户 token。
  version: "1.0.0"
servers:
  - url: https://your-go-file-host
    description: 替换为你的实际部署地址（如 http://1.2.3.4:3000）
security:
  - tokenAuth: []
paths:
  /api/ai/manifest:
    get:
      operationId: getManifest
      summary: 获取 API 自描述文档（公开，无需鉴权）
      security: []
      responses:
        "200":
          description: API 描述
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
  /api/ai/find:
    get:
      operationId: findFiles
      summary: 按文件名/关键词找文件（弱模型推荐先用）
      description: 返回最匹配的若干文件及一句自然语言 summary，便于模型决定下一步。
      parameters:
        - name: q
          in: query
          required: true
          description: 文件名或关键词
          schema: { type: string }
        - name: limit
          in: query
          description: 返回条数，默认 5，最大 20
          schema: { type: integer, minimum: 1, maximum: 20, default: 5 }
      responses:
        "200":
          description: 匹配结果
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items: { $ref: "#/components/schemas/AIFile" }
                          count: { type: integer }
                          summary: { type: string, description: 给模型的自然语言提示 }
  /api/ai/download:
    get:
      operationId: downloadByQuery
      summary: 按 id 或文件名一步下载（弱模型推荐）
      description: q 可为数字 id 或（部分）文件名，服务端解析最匹配项并返回其二进制内容。
      parameters:
        - name: q
          in: query
          required: true
          description: 文件 id 或文件名
          schema: { type: string }
      responses:
        "200":
          description: 文件二进制流
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        "404":
          description: 未找到匹配文件
  /api/ai/files:
    get:
      operationId: listFiles
      summary: 列出 / 搜索文件（分页）
      parameters:
        - name: q
          in: query
          description: 关键词，匹配文件名 / 描述 / 上传者 / 时间
          schema: { type: string }
        - name: page
          in: query
          description: 页码，默认 1
          schema: { type: integer, minimum: 1, default: 1 }
        - name: page_size
          in: query
          description: 每页数量，默认 20，最大 100
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: 文件分页列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items: { $ref: "#/components/schemas/AIFile" }
                          total: { type: integer }
                          page: { type: integer }
                          page_size: { type: integer }
        "403":
          description: 未鉴权或 token 无效
    post:
      operationId: uploadFile
      summary: 上传 / 回传文件
      description: multipart/form-data，字段 file 可重复以一次上传多个文件。
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: array
                  items: { type: string, format: binary }
                description:
                  type: string
                  description: 可选，附加到所有上传文件的描述
            encoding:
              file:
                contentType: application/octet-stream
      responses:
        "200":
          description: 上传成功，返回创建的文件记录
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items: { $ref: "#/components/schemas/AIFile" }
                          count: { type: integer }
        "413":
          description: 文件超过上传大小限制
  /api/ai/files/{id}:
    get:
      operationId: getFile
      summary: 获取单个文件元数据
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: 文件元数据
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/AIFile" }
        "404":
          description: 文件不存在
  /api/ai/files/{id}/content:
    get:
      operationId: downloadFile
      summary: 下载文件二进制内容
      description: 归档至冷存储的文件会自动取回后再返回。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: 文件二进制流
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        "404":
          description: 文件不存在
        "502":
          description: 归档文件取回失败
  /api/ai/stats:
    get:
      operationId: getStats
      summary: 文件维度统计
      responses:
        "200":
          description: 统计数据
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/FileStats" }
components:
  securitySchemes:
    tokenAuth:
      type: apiKey
      in: header
      name: Authorization
      description: 在用户设置页生成的 token，直接作为 Authorization 头的值（也兼容 "Bearer <token>"）。
  schemas:
    Envelope:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
        data: {}
    AIFile:
      type: object
      properties:
        id: { type: integer }
        filename: { type: string }
        description: { type: string }
        uploader: { type: string }
        time: { type: string }
        size: { type: integer, format: int64, description: 字节数 }
        size_human: { type: string, description: 人类可读大小，如 "5.00 MiB" }
        category: { type: string, description: 分类（图片/视频/音频/文档/压缩包/程序镜像/其他） }
        download_counter: { type: integer }
        archived: { type: boolean, description: 是否在冷存储 }
        download_url: { type: string, description: 可直接 GET（带 token）下载的绝对地址 }
    FileStats:
      type: object
      properties:
        total_files: { type: integer }
        total_size: { type: integer, format: int64 }
        total_downloads: { type: integer }
        archived_files: { type: integer }
        local_files: { type: integer }
        type_size:
          type: array
          items: { $ref: "#/components/schemas/NameValue" }
        type_count:
          type: array
          items: { $ref: "#/components/schemas/NameValue" }
        storage_state:
          type: array
          items: { $ref: "#/components/schemas/NameValue" }
        upload_trend:
          type: array
          items: { $ref: "#/components/schemas/NameValue" }
        top_downloads:
          type: array
          items: { $ref: "#/components/schemas/FileEntry" }
        largest_files:
          type: array
          items: { $ref: "#/components/schemas/FileEntry" }
    NameValue:
      type: object
      properties:
        name: { type: string }
        value: { type: integer, format: int64 }
    FileEntry:
      type: object
      properties:
        filename: { type: string }
        link: { type: string }
        uploader: { type: string }
        time: { type: string }
        size: { type: integer, format: int64 }
        download_counter: { type: integer }
        category: { type: string }
        archived: { type: boolean }